change tests to webdriver
